refactor 라우터 방식이 아닌 탭 방식 변경 + 컴포넌트 세분화
