+++ src/router/index.js
@@ -138,34 +138,8 @@ const routes = [
   // 태스크 관련
   {
     path: '/task/:taskId',
-    component: () => import('@/views/task/TaskDetail.vue'),
-    children: [
-      {
-        path: '',
-        name: 'TaskInfo',
-        component: () => import('@/components/task/TaskInfoTab.vue')
-      },
-      {
-        path: 'approval',
-        name: 'TaskApproval',
-        component: () => import('@/components/task/TaskApprovalTab.vue')
-      },
-      {
-        path: 'attachments',
-        name: 'TaskAttachment',
-        component: () => import('@/components/task/TaskAttachmentTab.vue')
-      },
-      {
-        path: 'comment',
-        name: 'TaskComment',
-        component: () => import('@/components/task/CommentTab.vue')
-      },
-      {
-        path: 'notice',
-        name: 'TaskNotice',
-        component: () => import('@/components/task/NoticeTab.vue')
-      }
-    ]
+    name: 'TaskDetail',
+    component: () => import('@/views/task/TaskDetail.vue')
   }
 ]
